+++ src/ch/epfl/leb/sass/simulator/generators/realtime/obstructors/commands/ObstructorCommandBuilder.java
@@ -15,23 +15,24 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.sass.simulator.generators.realtime.fluorophores.commands;
+package ch.epfl.leb.sass.simulator.generators.realtime.obstructors.commands;
 
 import ch.epfl.leb.sass.simulator.generators.realtime.components.Camera;
-import ch.epfl.leb.sass.simulator.generators.realtime.FluorophoreProperties;
+import ch.epfl.leb.sass.simulator.generators.realtime.components.Stage;
 import ch.epfl.leb.sass.simulator.generators.realtime.psfs.PSFBuilder;
 
 /**
- * Interface for populating the field with fluorophores.
+ * Interface for populating the field with obstructors, i.e. gold beads.
  * 
  * @author Kyle M. Douglass
  */
-public interface FluorophoreCommandBuilder {
+public interface ObstructorCommandBuilder {
     
-    public FluorophoreCommand build();
+    public ObstructorCommand build();
     
-    public FluorophoreCommandBuilder camera(Camera camera);
-    public FluorophoreCommandBuilder fluorProp(FluorophoreProperties fluorProp);
-    public FluorophoreCommandBuilder psfBuilder(PSFBuilder psfBuilder);
+    public ObstructorCommandBuilder camera(Camera camera);
+    public ObstructorCommandBuilder stage(Stage stage);
+    public ObstructorCommandBuilder brightness(double brightness);
+    public ObstructorCommandBuilder psfBuilder(PSFBuilder psfBuilder);
     
 }